requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
asyncpg==0.29.0
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
compressed_rtf==1.0.6
contourpy==1.2.1
cycler==0.12.1
databases==0.9.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.29.2
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
email_validator==2.1.1
executing==2.0.1
extract-msg==0.28.7
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.53.0
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.6
IMAPClient==2.1.0
importlib_metadata==8.0.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kiwisolver==1.4.5
langchain==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mysql==0.0.3
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.3
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
olefile==0.47
openai==0.28.0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pdfminer.six==20191110
peewee==3.17.5
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.22.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2024.5.15
requests==2.31.0
requests-futures==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
selenium==4.19.0
Send2Trash==1.8.2
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
streamlit==1.35.0
style==1.1.0
sympy==1.12.1
tbb==2021.13.0
tenacity==8.4.1
terminado==0.18.1
textract==1.6.5
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
transformers==4.41.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2023.4
tzlocal==5.2
ujson==5.10.0
update==0.0.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wordcloud==1.9.3
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.0
yahooquery==2.3.7
yarg==0.1.9
yarl==1.9.4
yfinance==0.2.40
zipp==3.19.2
